# book-network-ui/src/app/services/models/user-request.ts
/* tslint:disable */
/* eslint-disable */
export interface UserRequest {
  bio?: string;
  email: string;
  firstName: string;
  id?: number;
  lastName: string;
  password: string;
  profilePicture?: Array<string>;
}