# book-network-ui/src/app/services/models/user-response.ts
/* tslint:disable */
/* eslint-disable */
export interface UserResponse {
  bio?: string;
  createdAt?: string;
  email?: string;
  firstName?: string;
  id?: number;
  lastName?: string;
  profilePicture?: Array<string>;
  updatedAt?: string;
  updatedJWT?: string;
}